+++ server/src/main/java/org/opentosca/toscana/api/docs/CsarResources.java
@@ -1,8 +1,8 @@
-package org.opentosca.toscana.core.api.docs;
+package org.opentosca.toscana.api.docs;
 
 import java.util.List;
 
-import org.opentosca.toscana.core.api.model.CsarResponse;
+import org.opentosca.toscana.api.model.CsarResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
@@ -10,14 +10,14 @@ import io.swagger.annotations.ApiModelProperty;
 
 @ApiModel
 public class CsarResources extends HiddenResourceSupport {
-    
+
     @ApiModelProperty(required = true)
     @JsonProperty("_embedded")
     private CsarContainer embedded;
 
     @ApiModel
     private static class CsarContainer {
-        
+
         @ApiModelProperty(
             required = true,
             notes = "The list of csars stored on this system"
