+++ server/src/main/java/org/opentosca/toscana/api/docs/HiddenResources.java
@@ -1,4 +1,4 @@
-package org.opentosca.toscana.core.api.docs;
+package org.opentosca.toscana.api.docs;
 
 import java.util.List;
