+++ server/src/main/java/org/opentosca/toscana/api/docs/PlatformResources.java
@@ -1,8 +1,8 @@
-package org.opentosca.toscana.core.api.docs;
+package org.opentosca.toscana.api.docs;
 
 import java.util.List;
 
-import org.opentosca.toscana.core.api.model.PlatformResponse;
+import org.opentosca.toscana.api.model.PlatformResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
@@ -10,14 +10,14 @@ import io.swagger.annotations.ApiModelProperty;
 
 @ApiModel
 public class PlatformResources extends HiddenResourceSupport {
-    
+
     @ApiModelProperty(required = true)
     @JsonProperty("_embedded")
     private PlatformsContainer embedded;
 
     @ApiModel
     private static class PlatformsContainer {
-        
+
         @ApiModelProperty(
             required = true,
             notes = "The list of platforms supported by this system"
