+++ server/src/main/java/org/opentosca/toscana/api/docs/TransformationResources.java
@@ -1,8 +1,8 @@
-package org.opentosca.toscana.core.api.docs;
+package org.opentosca.toscana.api.docs;
 
 import java.util.List;
 
-import org.opentosca.toscana.core.api.model.TransformationResponse;
+import org.opentosca.toscana.api.model.TransformationResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
@@ -10,14 +10,14 @@ import io.swagger.annotations.ApiModelProperty;
 
 @ApiModel
 public class TransformationResources extends HiddenResourceSupport {
-    
+
     @ApiModelProperty(required = true)
     @JsonProperty("_embedded")
     private TransformationsContainer embedded;
 
     @ApiModel
     private static class TransformationsContainer {
-        
+
         @ApiModelProperty(
             required = true,
             notes = "The list of transformations for a csar"
