+++ server/src/main/java/org/opentosca/toscana/api/model/LogResponse.java
@@ -1,15 +1,16 @@
-package org.opentosca.toscana.core.api.model;
+package org.opentosca.toscana.api.model;
 
 import java.util.List;
 
-import org.opentosca.toscana.core.api.TransformationController;
-import org.opentosca.toscana.core.api.docs.HiddenResourceSupport;
+import org.opentosca.toscana.api.TransformationController;
+import org.opentosca.toscana.api.docs.HiddenResourceSupport;
 import org.opentosca.toscana.core.transformation.logging.LogEntry;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import org.springframework.hateoas.core.Relation;
+import org.springframework.hateoas.mvc.ControllerLinkBuilder;
 
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
@@ -32,7 +33,7 @@ public class LogResponse extends HiddenResourceSupport {
         this.end = end;
         this.logEntries = logEntries;
         //Add Self Link
-        this.add(linkTo(methodOn(TransformationController.class)
+        this.add(ControllerLinkBuilder.linkTo(methodOn(TransformationController.class)
             .getTransformationLogs(csarName, platform, start))
             .withSelfRel().expand(csarName));
         //Add link to retrieve only new log messages (next link)
