+++ server/src/test/java/org/opentosca/toscana/core/api/CsarControllerTest.java
@@ -9,6 +9,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Random;
 
+import org.opentosca.toscana.api.CsarController;
 import org.opentosca.toscana.core.BaseSpringTest;
 import org.opentosca.toscana.core.csar.Csar;
 import org.opentosca.toscana.core.csar.CsarImpl;
